Styling adjustments for traffic card, restyling for UsersTable
